requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
pandas
langchain-groq
langchain
langgraph
openpyxl
matplotlib